rename module, handle asyncness.
